refactor: 컴포넌트 폴더별로 정리
- 사용되는 View에 따라 폴더를 생성하여 컴포넌트 위치
