+++ src/view/LoginView.ts
@@ -68,7 +68,7 @@ export class LoginView implements LoginViewInterface {
   #setUserInfo(userInfo) {
     const { accessToken, user } = userInfo;
 
-    localStorage.setItem('accessToken', JSON.stringify(accessToken));
+    localStorage.setItem('accessToken', accessToken);
     localStorage.setItem('user', JSON.stringify(user));
   }
 }
